% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\docType{data}
\name{platemap}
\alias{platemap}
\title{Plate layout.}
\format{
A data frame with 308 rows and 4 variables:
\describe{
\item{Row}{the row number of the well}
\item{Column}{the column number of the well}
\item{gene}{what gene is perturbed in the well}
\item{compound}{what compound is added to the well}
}
}
\usage{
platemap
}
\description{
Plate layout made by Ezgi Tanriver Ayder.
}
\keyword{datasets}
