% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single_plate.R
\name{sim_single_plate}
\alias{sim_single_plate}
\title{Simulate a single plate}
\usage{
sim_single_plate(
  plate_layout,
  hit_genes,
  hit_genes_with_interaction,
  desired_cells_per_well,
  base_level,
  mu_bg,
  sigma_bg,
  mu_btg,
  sigma_btg,
  cv_cell,
  measure_error,
  spatial_effect,
  sigma_bg0
)
}
\arguments{
\item{plate_layout}{data frame similar to platemap}

\item{hit_genes}{number of hit genes}

\item{hit_genes_with_interaction}{number of hit genes
with interaction with compound}

\item{desired_cells_per_well}{numeric value desired number of cells per well}

\item{base_level}{numeric value baseline expression of the target gene}

\item{mu_bg}{average gene effect on expression, percentage}

\item{sigma_bg}{between gene variation in effect on target expression}

\item{mu_btg}{average gene treatment interaction, percentage}

\item{sigma_btg}{between gene variation in interaction}

\item{cv_cell}{between cell coefficient of variation}

\item{measure_error}{measurement error log scale}

\item{spatial_effect}{data frame wells with three columns, Row, Column
and spatial bias size}

\item{sigma_bg0}{between gene variation among the true negative gene editing}
}
\value{
data frame
}
\description{
Simulate a single plate
}
