% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells_with_spatial_bias.R
\docType{data}
\name{wells_with_spatial_bias}
\alias{wells_with_spatial_bias}
\title{Example wells that are affected by spatial bias.}
\format{
A data frame with 48 rows and 3 variables:
\describe{
\item{Row}{the row number of the well}
\item{Column}{the column number of the well}
\item{spatial_bias}{size of the spatial bias}
}
}
\usage{
wells_with_spatial_bias
}
\description{
Example wells that are affected by spatial bias.
}
\keyword{datasets}
